# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogueView": {
        "properties": {
          "deprecated": {
            "description": "Set when the module is no longer offered. A picker should hide it.",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "name",
          "requires"
        ],
        "type": "object"
      },
...
      "Credentials": {
        "example": {
          "handle": "owner@acme.example",
          "password": "correct horse battery staple"
        },
        "properties": {
          "handle": {
            "description": "The email address the account was registered with.",
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "handle",
          "password"
        ],
        "type": "object"
      },
...
      "Problem": {
        "example": {
          "code": "auth.invalid_credentials",
          "detail": "Those login details are not right. Please try again.",
          "status": 401,
          "title": "Unauthorized",
          "type": "https://errors.spa.example/auth.invalid_credentials"
        },
        "properties": {
          "args": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MessageArg"
            },
            "description": "The values the message names, typed — which account, which module, how\nmuch was outstanding. A client that renders its own sentence reads these\ninstead of parsing `detail`. Absent when the message takes none.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "code": {
            "description": "The stable identifier. Branch on this, never on `detail`.",
            "type": "string"
          },
          "detail": {
            "description": "The message in the caller's language, chosen by `Accept-Language`.",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "description": "`https://errors.spa.example/` followed by the code. Stable, and a place\nto look it up — see `docs/ERRORS.md` for every code this API can answer\nwith.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "type": "object"
      },
...
      "SessionCreated": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "token": {
            "description": "Send as `Authorization: Bearer <token>`.",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ],
        "type": "object"
      },
...
      "TenantView": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "modules": {
            "description": "The modules this tenant has turned on. A route belonging to any other\nmodule answers 404 here.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "role": {
            "description": "What the caller may do here, so a client can hide what it must not\noffer. The server refuses regardless — this is for the buttons.\n\nOne of `owner`, `accountant`, `clerk`, `viewer`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "modules"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "A token from `POST /v1/sessions`. Sessions expire; a 401 with `auth.session_expired` means log in again.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "SPA",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/catalogue": {
      "get": {
        "description": "Unauthenticated, and on the **apex** — a signup form and a pricing page both\nneed this before anyone has an account, and before there is a subdomain to\nask from. It is product information, not data.\n\n`/v1/catalogue` rather than `/v1/modules` because that is now what a tenant's\nown list is called. The two collided the moment the tenant moved to the\nsubdomain, and the router refused to start — correctly: a path has to name\none thing, because nothing routes on the host.",
        "operationId": "catalogue",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogueView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Done."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "security": [],
        "summary": "Every module this build offers.",
        "tags": [
          "modules"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ]
    },
...
    "/v1/openapi.json": {
      "get": {
        "description": "Point any `OpenAPI` viewer at it. Deliberately not a bundled UI: the bundles\nare megabytes of vendored assets fetched at build time, which is a network\ndependency in a build that otherwise has none.",
        "operationId": "openapi_json",
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "This document, as OpenAPI 3.1"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "security": [],
        "summary": "This document.",
        "tags": [
          "service"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ]
    },
...
    "/v1/sessions": {
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "log_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body is not JSON. `code` is `request.malformed_body`, and `args.reason` says what the parser found."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Wrong handle or password — the same answer for both, deliberately"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body was sent without `Content-Type: application/json`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body is JSON and not the shape this route takes. `args.reason` names the field."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "security": [],
        "summary": "Log in.",
        "tags": [
          "sessions"
        ]
      }
    },
    "/v1/sessions/current": {
      "delete": {
        "operationId": "log_out",
        "responses": {
          "204": {
            "description": "Ended. The token no longer works."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "summary": "Log out, ending the session this request authenticated with.",
        "tags": [
          "sessions"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ]
    },
...
    "/v1/tenant": {
      "get": {
        "description": "Proves the whole path: `Allowed<Read>` in the signature means membership was\nchecked *and* the role permits it, and there is no way to obtain one that\nskips either.",
        "operationId": "tenant",
        "parameters": [
          {
            "description": "The tenant's subdomain — `bassat.spa.com`. Every path below is about that tenant.",
            "in": "header",
            "name": "Host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantView"
                }
              }
            },
            "description": "Done."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such tenant, or not one of yours — the same answer for both, so the API is not a tenant-enumeration oracle"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The tenant is real and entitled, and not serving right now. Retryable."
          }
        },
        "summary": "What this tenant is, and what you may do in it.",
        "tags": [
          "tenants"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ]
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}